feat: pagination in NewsGridContainer #696
moved pagination code into the component. used if to only show the buttons if there is a next page
